refactor: rename internal/gateway/client -> internal/cloudclient
The package is the hosted Cloud API side-channel client (websearch/webfetch/
advisor/BYOK over /v1/*), not a gateway runtime. Reserve internal/gateway/
for the real self-hostable gateway. Updates the two importers
(provider/byok.go, provider/wire.go) and the guard-test path assertion; no
behavior change, no new dependencies.
