+++ internal/cloudclient/web.go
@@ -1,4 +1,4 @@
-package client
+package cloudclient
 
 import (
 	"context"
